+++ emperor/support_files/js/controller.js
@@ -28,8 +28,8 @@ define([
 
   var EmperorAttributeABC = viewcontroller.EmperorAttributeABC;
 
-  TAB_ORDER = ['color', 'visibility', 'opacity', 'scale',
-               'shape', 'axes', 'animations'];
+  var TAB_ORDER = ['color', 'visibility', 'opacity', 'scale',
+                   'shape', 'axes', 'animations'];
 
   var controllerConstructors = {
       'color': ColorViewController,
@@ -626,7 +626,7 @@ define([
       if (item === 'shape' && isLargeDataset)
         continue;
       scope.controllers[item] = scope.addTab(scope.sceneViews[0].decViews,
-                                           controllerConstructors[item]);
+                                             controllerConstructors[item]);
     }
 
     // We are tabifying this div, I don't know man.